apply self code review
