+++ pkg/registry/common/dial/ns_client.go
@@ -159,7 +159,9 @@ func (c *dialNSClient) Find(ctx context.Context, in *registry.NetworkServiceQuer
 
 // NewNetworkServiceRegistryClient - returns a new null client that does nothing but call next.NetworkServiceRegistryClient(ctx).
 func NewNetworkServiceRegistryClient(chainCtx context.Context, opts ...Option) registry.NetworkServiceRegistryClient {
-	o := &option{}
+	o := &option{
+		dialTimeout: time.Millisecond * 100,
+	}
 	for _, opt := range opts {
 		opt(o)
 	}
